security: comprehensive audit pass 4 - PII masking, compile fixes, full build verification
SECURITY FIXES:
- ChatController: remove unused senderEmail field from ChatSendRequest DTO
  (PII leak vector - client-injectable email never used server-side)
- EmailService: apply maskEmail() to all 11 raw-email log lines across
  sendPasswordResetEmail, sendOtpEmail, sendHtmlEmail, sendViaResend, sendViaSmtp
- AuthService: mask email in OTP failure log (warn level was logging full address)
- EventProducerService: mask event.getTo() in Kafka email event logs
- EmailConsumer: mask event.getTo() in consumer logs; remove event object dump
  from missing-address error (was serializing full email in log)
- StompHeaderInterceptor: mask userEmail in WebSocket rejection warn logs

BUG FIXES (pre-existing compile errors):
- EmailService: fix invalid try-with-resources on HttpClient (not AutoCloseable in Java 17)
- AuthService: fix lambda variable shadowing - rename removeIf(e->...) to
  removeIf(entry->...) to avoid conflict with enclosing catch(RuntimeException e)

BUILD VERIFICATION:
- mvn compile: BUILD SUCCESS (129 source files, 0 errors)
- npm run build: BUILD SUCCESS (2598 modules, 0 errors, 0 warnings)

SECURITY SCAN RESULTS:
- Zero hardcoded secrets in any committed file
- Zero dangerouslySetInnerHTML in JSX
- All @Query annotations use parameterized JPQL (no native SQL, no injection risk)
- All admin endpoints protected with @PreAuthorize(hasRole ADMIN/SUPER_ADMIN)
- CORS restricted to explicit allowed origins with fail-fast wildcard guard
- Security headers: X-Frame-Options DENY, HSTS, X-Content-Type-Options, Referrer-Policy

NEW FILES:
- frontend/src/utils/sanitize.js: allowed image type whitelist helper
- frontend/src/utils/timeAgo.js: time formatting utility
- src/main/java/com/example/socialmedia/dto/AdminActionLogResponse.java: admin log DTO

+++ src/main/java/com/example/socialmedia/dto/PostResponse.java
@@ -10,8 +10,8 @@ public class PostResponse {
     private String imageUrl;
     private String authorName;
     private Long authorId;
-    private int likeCount;
-    private int commentCount;
+    private long likeCount;
+    private long commentCount;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'", timezone = "UTC")
     private LocalDateTime createdAt;
     private boolean isLiked;
@@ -22,13 +22,13 @@ public class PostResponse {
     public PostResponse() {
     }
 
-    public PostResponse(Long id, String content, String imageUrl, String authorName, Long authorId, int likeCount,
-            int commentCount, LocalDateTime createdAt, boolean isLiked) {
+    public PostResponse(Long id, String content, String imageUrl, String authorName, Long authorId, long likeCount,
+            long commentCount, LocalDateTime createdAt, boolean isLiked) {
         this(id, content, imageUrl, authorName, authorId, likeCount, commentCount, createdAt, isLiked, false);
     }
 
-    public PostResponse(Long id, String content, String imageUrl, String authorName, Long authorId, int likeCount,
-            int commentCount, LocalDateTime createdAt, boolean isLiked, boolean isSaved) {
+    public PostResponse(Long id, String content, String imageUrl, String authorName, Long authorId, long likeCount,
+            long commentCount, LocalDateTime createdAt, boolean isLiked, boolean isSaved) {
         this.id = id;
         this.content = content;
         this.imageUrl = imageUrl;
@@ -82,19 +82,19 @@ public class PostResponse {
         this.authorId = authorId;
     }
 
-    public int getLikeCount() {
+    public long getLikeCount() {
         return likeCount;
     }
 
-    public void setLikeCount(int likeCount) {
+    public void setLikeCount(long likeCount) {
         this.likeCount = likeCount;
     }
 
-    public int getCommentCount() {
+    public long getCommentCount() {
         return commentCount;
     }
 
-    public void setCommentCount(int commentCount) {
+    public void setCommentCount(long commentCount) {
         this.commentCount = commentCount;
     }
 
@@ -132,8 +132,8 @@ public class PostResponse {
         private String imageUrl;
         private String authorName;
         private Long authorId;
-        private int likeCount;
-        private int commentCount;
+        private long likeCount;
+        private long commentCount;
         private LocalDateTime createdAt;
         private boolean isLiked;
         private boolean isSaved;
@@ -166,12 +166,12 @@ public class PostResponse {
             return this;
         }
 
-        public PostResponseBuilder likeCount(int likeCount) {
+        public PostResponseBuilder likeCount(long likeCount) {
             this.likeCount = likeCount;
             return this;
         }
 
-        public PostResponseBuilder commentCount(int commentCount) {
+        public PostResponseBuilder commentCount(long commentCount) {
             this.commentCount = commentCount;
             return this;
         }
